+++ drivers/iio/adc/cf_axi_adc.h
@@ -11,7 +11,7 @@
 #ifndef ADI_AXI_ADC_H_
 #define ADI_AXI_ADC_H_
 
-#include <linux/fpga/adi-axi-common.h>
+#include <linux/adi-axi-common.h>
 #include <linux/iio/iio.h>
 
 /* ADC COMMON */
