+++ drivers/iio/adc/cf_axi_tdd.c
@@ -5,11 +5,11 @@
  * Copyright 2016 Analog Devices Inc.
  *
  */
+#include <linux/adi-axi-common.h>
 #include <linux/bitfield.h>
 #include <linux/bits.h>
 #include <linux/clk.h>
 #include <linux/device.h>
-#include <linux/fpga/adi-axi-common.h>
 #include <linux/iio/iio.h>
 #include <linux/io.h>
 #include <linux/math64.h>
