+++ drivers/media/platform/adi/adi-axi-fb.c
@@ -5,12 +5,12 @@
  *
  */
 
+#include <linux/adi-axi-common.h>
 #include <linux/device.h>
 #include <linux/module.h>
 #include <linux/of_address.h>
 #include <linux/platform_device.h>
 #include <media/media-device.h>
-#include <linux/fpga/adi-axi-common.h>
 
 /* DMA defines */
 #define DMAC_REG_CTRL				0x400
