+++ drivers/misc/adi-axi-tdd.c
@@ -5,11 +5,11 @@
  * Copyright 2023 Analog Devices Inc.
  *
  */
+#include <linux/adi-axi-common.h>
 #include <linux/bitfield.h>
 #include <linux/bits.h>
 #include <linux/clk.h>
 #include <linux/device.h>
-#include <linux/fpga/adi-axi-common.h>
 #include <linux/mod_devicetable.h>
 #include <linux/module.h>
 #include <linux/notifier.h>
