+++ include/linux/adi-axi-common.h
@@ -14,7 +14,6 @@
 #define ADI_AXI_REG_VERSION			0x0000
 #define ADI_AXI_REG_ID				0x0004
 #define ADI_AXI_REG_SCRATCH			0x0008
-#define ADI_AXI_REG_CONFIG			0x000c
 
 #define ADI_AXI_REG_FPGA_INFO			0x001C
 #define ADI_AXI_REG_FPGA_VOLTAGE		0x0140
